Refactor insertCode method to use Integer type and update related logic in AuthService and AuthServiceImpl

+++ src/main/java/pt/gov/chavemoveldigital/controllers/AuthController.java
@@ -2,7 +2,6 @@ package pt.gov.chavemoveldigital.controllers;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,7 +9,6 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import pt.gov.chavemoveldigital.entities.TempCode;
 import pt.gov.chavemoveldigital.models.UserDTO;
 import pt.gov.chavemoveldigital.services.AuthService;
-import org.springframework.ui.Model;
 
 
 @Controller
@@ -28,7 +26,7 @@ public class AuthController {
     }
 
     @PostMapping("/authentication/code")
-    public ResponseEntity<?> insertCode(Long code) {
+    public ResponseEntity<?> insertCode(Integer code) {
         return ResponseEntity.ok().body(authService.insertCode(code));
     }
 }
